feat: applicationList, fix span startTime etc.
- renmae serviceList -> applicationlist, and support Duration as parameter
- add /trace/options route
- fix span's startTime out of graphql.Int( use graphsql.Float instead )
- support oprationName and applicationId(service name) as /trace filters
- cmd/query/graphql/README.md describe graphql interface

+++ cmd/query/app/graphql/span.go
@@ -80,7 +80,7 @@ var GLSpanType = graphql.NewObject(
 				Type: graphql.NewList(GLReferenceType),
 			},
 			"startTime": &graphql.Field{
-				Type: graphql.Int,
+				Type: graphql.Float,
 			},
 			"duration": &graphql.Field{
 				Type: graphql.Int,
